# Optimized std.datetime.date.DateTime from string methods