In the vein of https://github.com/dlang/phobos/pull/5413 and https://github.com/dlang/phobos/pull/5404