@JackStouffer 2017-07-06
Speed results

```
# old
$ git checkout master
$ dmd/src/dmd -O -inline -release test.d && ./test
result	6 secs, 83 ms, 931 μs, and 1 hnsec

# new
$ git checkout DateTime-fromstring
$ dmd/src/dmd -O -inline -release test.d && ./test
result	1 sec, 746 ms, and 330 μs
```

code 

```d
import std.stdio;
import std.algorithm;
import std.conv;
import std.format;
import std.ascii;
import std.traits;
import std.datetime;

enum testCount = 5_000_000;


void main()
{
    auto result = to!Duration(benchmark!(() => DateTime.fromISOString("20100704T070612"))(testCount)[0]);

    writeln("result", "\t", result);
}
```